refactor(discord): move configuration type into discord package
