+++ notify/discord/discord_test.go
@@ -31,7 +31,6 @@ import (
 
 	amcommoncfg "github.com/prometheus/alertmanager/config/common"
 
-	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/notify"
 	"github.com/prometheus/alertmanager/notify/test"
 	"github.com/prometheus/alertmanager/types"
@@ -42,7 +41,7 @@ var testWebhookURL, _ = url.Parse("https://discord.com/api/webhooks/971139602272
 
 func TestDiscordRetry(t *testing.T) {
 	notifier, err := New(
-		&config.DiscordConfig{
+		&DiscordConfig{
 			WebhookURL: &amcommoncfg.SecretURL{URL: testWebhookURL},
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -71,14 +70,14 @@ func TestDiscordTemplating(t *testing.T) {
 
 	for _, tc := range []struct {
 		title string
-		cfg   *config.DiscordConfig
+		cfg   *DiscordConfig
 
 		retry  bool
 		errMsg string
 	}{
 		{
 			title: "full-blown message",
-			cfg: &config.DiscordConfig{
+			cfg: &DiscordConfig{
 				Title:   `{{ template "discord.default.title" . }}`,
 				Message: `{{ template "discord.default.message" . }}`,
 			},
@@ -86,14 +85,14 @@ func TestDiscordTemplating(t *testing.T) {
 		},
 		{
 			title: "title with templating errors",
-			cfg: &config.DiscordConfig{
+			cfg: &DiscordConfig{
 				Title: "{{ ",
 			},
 			errMsg: "template: :1: unclosed action",
 		},
 		{
 			title: "message with templating errors",
-			cfg: &config.DiscordConfig{
+			cfg: &DiscordConfig{
 				Title:   `{{ template "discord.default.title" . }}`,
 				Message: "{{ ",
 			},
@@ -137,7 +136,7 @@ func TestDiscordRedactedURL(t *testing.T) {
 
 	secret := "secret"
 	notifier, err := New(
-		&config.DiscordConfig{
+		&DiscordConfig{
 			WebhookURL: &amcommoncfg.SecretURL{URL: u},
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -159,7 +158,7 @@ func TestDiscordReadingURLFromFile(t *testing.T) {
 	require.NoError(t, err, "writing to temp file failed")
 
 	notifier, err := New(
-		&config.DiscordConfig{
+		&DiscordConfig{
 			WebhookURLFile: f.Name(),
 			HTTPConfig:     &commoncfg.HTTPClientConfig{},
 		},
@@ -193,7 +192,7 @@ func TestDiscord_Notify(t *testing.T) {
 	require.NoError(t, err)
 
 	// Create a DiscordConfig with the WebhookURLFile set
-	cfg := &config.DiscordConfig{
+	cfg := &DiscordConfig{
 		WebhookURLFile: tempFile.Name(),
 		HTTPConfig:     &commoncfg.HTTPClientConfig{},
 		Title:          "Test Title",
